Don't pass Unicode strings to File::Spec->rel2abs()
Otherwise the resulting filename has the non-ASCII characters
double-encoded in UTF-8 (and will not work). Filenames don't have
any associated character set and should thus always be treated as bytes.

See https://bugzilla.redhat.com/show_bug.cgi?id=1171670 for a discussion
of this problem in the context of Publican.

Signed-off-by: Raphaël Hertzog <hertzog@debian.org>